requirements.txt
numpy==1.22.2
flask>=2.0.0
python-dotenv>=0.6.3
pydub==0.21.0
safetensors==0.3.1
setuptools
openai-whisper